test other composable for Calendar picker; move extension function to common StringToDateExtension.kt
