+++ app/src/main/res/values/strings.xml
@@ -22,4 +22,5 @@
     <string name="sign_up">SIGN UP</string>
     <string name="invalid_or_missing_token_or_api_key">Invalid or missing token or API key</string>
     <string name="logout">Logout</string>
+    <string name="cancel">CANCEL</string>
 </resources>
